Imported fixes to ClosingStructureEntity
Related to Issue [WTI-905]


git-svn-id: https://repos.deltares.nl/repos/WettelijkToetsInstrumentarium/trunk@6132 5a3db67b-de53-47b5-99c8-a1c30a6650e2

Former-commit-id: 670466348426e6c72f840259ae707f003587e406
Former-commit-id: f93a9fc7979a457f9fd718c7d71675d3b7fec4e2

+++ Application/Ringtoets/src/Application.Ringtoets.Storage/Read/ClosingStructures/ClosingStructureEntityReadExtensions.cs
@@ -105,7 +105,7 @@ namespace Application.Ringtoets.Storage.Read.ClosingStructures
                 },
                 ProbabilityOpenStructureBeforeFlooding = entity.ProbabilityOpenStructureBeforeFlooding.ToNullAsNaN(),
                 FailureProbabilityOpenStructure = entity.FailureProbabilityOpenStructure.ToNullAsNaN(),
-                IdenticalApertures = (int) entity.IdenticalApertures,
+                IdenticalApertures = entity.IdenticalApertures,
                 FailureProbabilityReparation = entity.FailureProbabilityReparation.ToNullAsNaN(),
                 InflowModelType = (ClosingStructureInflowModelType) entity.InflowModelType,
             });
